backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.7.0
google-genai
python-multipart
python-dotenv
httpx
langgraph
numpy
fpdf2
firebase-admin
pypdf
python-docx
gunicorn
